chore(map): refactor functions
